+++ src/aux-server/aux-backend/shared/webhooks/SimulationWebhookEnvironment.ts
@@ -194,6 +194,7 @@ export class SimulationWebhookEnvironment implements WebhookEnvironment {
                         websocketProtocol,
                     };
                 },
+                () => null,
                 undefined
             );