+++ src/aux-vm-browser/managers/BotManager.ts
@@ -374,6 +374,7 @@ export class BotManager extends BaseSimulation implements BrowserSimulation {
                     endpoint,
                     authenticateIfNotLoggedIn
                 ),
+            (endpoint) => this._getAuthEndpointHelper(endpoint),
             undefined,
             (endpoint, protocol) => {
                 if (protocol === 'apiary-aws') {